% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyState.R
\name{addPloidySvmLineToPlot}
\alias{addPloidySvmLineToPlot}
\title{add a separation line to a ploidy vs 2N+LOH plot}
\usage{
addPloidySvmLineToPlot(col = "gray", lty = "solid", lwd = 1.5)
}
\arguments{
\item{col}{line color}

\item{lty}{line type}

\item{lwd}{line width}
}
\description{
separate near-diploid from high-ploidy
}
