% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationUtils.R
\name{allelicCNV}
\alias{allelicCNV}
\title{allele specific cnv}
\usage{
allelicCNV(starLookUp, segmentData)
}
\arguments{
\item{starLookUp}{object returned by \code{makeStarLookUpTable}}

\item{segmentData}{segments from segmentation that are longer than minReasonableSegmentSize and then broken up into segments no smaller than 3 Mbs}
}
\description{
\enumerate{
\item find CN for the segment by rounding the cnLevel to an integer
\item find which closest star, within that CN, is closest to the hetScore of the segment
\item the closest star is then the segment's major:minor allele
}
}
\keyword{internal}
