% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{assignGridHeights}
\alias{assignGridHeights}
\title{each peak gets a group of grid heights based on the mode of the peak}
\usage{
assignGridHeights(peakInfo, n00, minGridHeight)
}
\arguments{
\item{peakInfo}{summary table of info for each peak found in \code{peaksByDensity}}

\item{n00}{Precision for creating initial grid}

\item{minGridHeight}{minimum value that can be assigned to the gridHeights}
}
\description{
each peak gets a group of grid heights based on the mode of the peak
}
