% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{binnedPosEnd}
\alias{binnedPosEnd}
\title{For given position and bin size, return the binned position}
\usage{
binnedPosEnd(pos, binSize = 1000)
}
\arguments{
\item{pos}{Vector of positions}

\item{binSize}{Size of the bin}
}
\value{
The bin just before the bin containing the ending position (last fully populated bin)
}
\description{
This is done for consistency. There are many off by one errors
one can make when doing this.
}
\details{
This function assumes the following bining for
1K and 10K bins and will return the \strong{bin ends}:

\tabular{ccc}{
bin# \tab 1K \tab 10K \cr
1 \tab    1..\strong{1000} \tab     1..\strong{10000} \cr
2 \tab 1001..\strong{2000} \tab 10001..\strong{20000} \cr
3 \tab 2001..\strong{3000} \tab 20001..\strong{30000} \cr
\dots \tab \tab \cr
}
}
\seealso{
Other coordinates: 
\code{\link{bimaToLinear}()},
\code{\link{binnedPosStart}()},
\code{\link{getLinearCoordinates}()},
\code{\link{linearCoordinates}}
}
\concept{coordinates}
