% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{binnedPosStart}
\alias{binnedPosStart}
\title{For given position and bin size, return the binned position}
\usage{
binnedPosStart(pos, binSize = 1000)
}
\arguments{
\item{pos}{Vector of positions}

\item{binSize}{Size of the bin}
}
\value{
The bin which contains the starting position.
}
\description{
This is done for consistency. There are many off by one errors
one can make when doing this.
}
\details{
This function assumes the following binning for
1K and 10K bins and will return the \strong{bin starts}:

\tabular{ccc}{
bin# \tab 1K \tab 10K \cr
1 \tab    \strong{1}..1000 \tab     \strong{1}..10000 \cr
2 \tab \strong{1001}..2000 \tab \strong{10001}..20000 \cr
3 \tab \strong{2001}..3000 \tab \strong{20001}..30000 \cr
\dots \tab \tab \cr
}
}
\seealso{
Other coordinates: 
\code{\link{binnedPosEnd}()},
\code{\link{getLinearCoordinates}()},
\code{\link{linearCoordinates}}
}
\concept{coordinates}
