% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{calcCopyNumber}
\alias{calcCopyNumber}
\title{find copy number from read depth and tumor ratio}
\usage{
calcCopyNumber(NRD, tau)
}
\arguments{
\item{NRD}{normalized read depth}

\item{tau}{tumor ratio}
}
\description{
find copy number from read depth and tumor ratio
}
\examples{
calcCopyNumber(NRD=23,tau=0.85) # CCND1 PT140
calcCopyNumber(NRD=1.4,tau=0.33) # CDKN2A BL002
calcCopyNumber(NRD=4,tau=0.75)
calcCopyNumber(NRD=4,tau=0.5)
calcCopyNumber(NRD=1.745,tau=0.13)
}
\seealso{
Other copyNumberCalcs: 
\code{\link{calcNrd}()},
\code{\link{calcTau}()},
\code{\link{calcTumorRatio}()}
}
\concept{copyNumberCalcs}
