% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{calcTumorFromPloidyPeaks}
\alias{calcTumorFromPloidyPeaks}
\title{calculate tumor from ploidy read depth and copy number calls (peak positions)}
\usage{
calcTumorFromPloidyPeaks(peakCopyNum, peakHeight, peakReadDepth_1bp, dPeaks)
}
\arguments{
\item{peakCopyNum}{from peakInfo, copy number for the peak}

\item{peakHeight}{from peakInfo, height of each peak normalized so max peak = 1, to determine two best peaks}

\item{peakReadDepth_1bp}{from peakInfo, read depth per peak normalized to 1bp window}
}
\description{
use \code{getTwoBestPeakIndexes} to find best peak and the biggest of its neighbor peaks.
}
