% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{calculateHetScore}
\alias{calculateHetScore}
\title{Combines ref/alt SNP data from each chromosome 1-22,X}
\usage{
calculateHetScore(
  sampleId,
  alternateId = NULL,
  inputDir,
  outputDir,
  segmentation = NULL,
  noPdf = FALSE,
  maximumCoverage = 1000,
  trimFromAlt = 2,
  trimFromRef = 1,
  trimExtraPerCoverage = 0.1,
  minSnpsToCalculateStatistic = 20,
  samplingStep = 30000,
  extraWindow = 1e+06,
  readDepthPer30kbBin = NULL,
  readDepthBinSize = 30000,
  gainColor = "blue",
  lossColor = "red",
  noDelGainCalls = FALSE
)
}
\arguments{
\item{sampleId}{sample identifier, used as prefix for all input and output file names
\code{<sampleId>_snpVals_<chromosome>.Rds} and \code{<sampleId>_countBP_<chromosome>.Rds}}

\item{alternateId}{optional secondary sample identifier}

\item{inputDir}{full path to snpVal and countBP Rds, segmentation data}

\item{outputDir}{full path to directory to output all files and intermediate objects created}

\item{segmentation}{identified regions of the genome with constant read depth. Data.frame with required columns:
chr, start, end, rd; optional: cnvState (1=loss, 2=normal, 3=gain) for color coded linear linear genome plot}

\item{noPdf}{if present (TRUE), pdf files will not be generated, instead plots are drawn on default device}

\item{maximumCoverage}{Do not process SNPs covered more than this}

\item{trimFromAlt}{Bins to discard from the 'alt' side of the distribution}

\item{trimFromRef}{Bins to discard from the 'ref' side of the distribution}

\item{trimExtraPerCoverage}{Fraction of bins to trim per each extra coverage}

\item{minSnpsToCalculateStatistic}{Minimum SNPs required in the window (1Mb) to calculate the statistic}

\item{samplingStep}{How frequently to try to summarize the data (bp), to produce overlapping windows}

\item{extraWindow}{Size of the window, how many bps to look at during each sampling step}

\item{readDepthPer30kbBin}{list with two equal length arrays:
\itemize{
\item \code{readDepthArray} read depth for 30 kb bins, normalized for GC content and other artifacts
\item \code{goodWindowArray} linear genome position of each window (bin), masked windows have been removed
}}

\item{readDepthBinSize}{bp size of the bins in the read depth array, expecting 30000 or 100000}

\item{gainColor}{color to use for gains in linear genome plot, default is blue}

\item{lossColor}{color to use for losses in linear geneome plot, default is red}

\item{noDelGainCalls}{do not color code deletions and gains in genome plot}
}
\description{
Calculate heterozygosity score by bin (.wig) and by chromosome arm (.csv), save to outputDir.
}
\details{
Will automatically create the the three panel plot, either to pdf or plot window.
This three panel plot can also be called separately, see \link{makeHetScoreReportPdf}
}
\examples{
\dontrun{

  library(BACDAC)
  library(logging)
  # calculateHetScoreExample.R
  # generate heterozygosity score by bin and by chromosome arm and save to the output directory

  # make a 3-panel plot, top panel is the segmentation data and the next two panels are the
  # resulting hetscore by bin and by arm

  basicConfig("DEBUG")
  sampleId='TCGA-14-1402-02A_ds'; alternateId=66301
  outputDir <- tempdir()

  # inputDir is the path to the load package data
  inputDir <- system.file('extdata', package = "BACDAC")
  segmentationFile <- file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
  segmentation <- read.csv(segmentationFile, comment.char = '#') # chr, start, end, rd per
  segmentation <- checkSegmentation(segmentation)

  thirtyKbFile=file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds'))
  readDepthPer30kbBin = readRDS(file=thirtyKbFile )

  calculateHetScore(
    sampleId=sampleId,
    inputDir=inputDir,
    outputDir=outputDir,
    segmentation=segmentation,
    noPdf = TRUE,
    #optional - used for the top panel of the three panel plot
    readDepthPer30kbBin=readDepthPer30kbBin,
    readDepthBinSize=readDepthPer30kbBin$windowSize
   )
}
}
