% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{calculateHetScoreTestStatisticPerCoverage}
\alias{calculateHetScoreTestStatisticPerCoverage}
\title{Calculate the value of test statistic for HetScore analysis}
\usage{
calculateHetScoreTestStatisticPerCoverage(
  maximumCoverage,
  trimFromAlt = 2,
  trimFromRef = 1,
  trimExtraPerCoverage = 0.1
)
}
\arguments{
\item{maximumCoverage}{We return an array with test statistics for coverage going
from 1 to maximumCoverage}

\item{trimFromAlt}{How many bins to ignore at the "alt" side}

\item{trimFromRef}{How many bins to ignore at the "ref" side}

\item{trimExtraPerCoverage}{Fraction of bins to trim for coverage.
If coverage == 10 and \code{trimExtraPerCoverage} is 0.1, we trim one extra bin}
}
\value{
Array with test statistic for each coverage from 1 to maximumCoverage
}
\description{
The statistic is expected number of times we observe reference/alternate
allele (whichever count is smaller) for given coverage, assuming there is no
loss of heterozygosity, so the ref/alt observations are of equal probability.
}
\details{
Let's say coverage is 4. We can observe 16 different combinations of
ref/alt on these 4 reads. These will contain:
1 times - 0 ref alleles << problem
4 times - 1 ref allele  << problem
6 times - 2 ref alleles = 2 alt alleles
4 times - 3 ref alleles
1 times - 4 ref alleles << problem

Now, there is a problem with 0/1 ref allele bins. For each SNP location we look at,
we have no idea if the person in question is actually heterozygous. So many times
we would encounter 0 alt / 0 ref situations. Since we do not know how many
SNPs the sample is heterozygous four, we cannot use these numbers.
Moreover, it appears that the bin #1 (1 ref allele) tends to get 'tainted'
because just by sheer luck sometimes you measure a wrong base pair even though
the sample is homozygous. So the values "spill" from bin #0 into bin #1.
Similar situation is on the other end (not symmetrical- due to mismappings on the left side)
Thus we define the \code{trimFromAlt} and \code{trimFromRef} values that will
let us skip calculating the test statistic if the bin is a problematic one.
}
\seealso{
Other hetScore: 
\code{\link{makeAndSaveHetScorePerArm}()},
\code{\link{plotHetScorePerArm}()},
\code{\link{plotHetScorePerBin}()},
\code{\link{saveHetScoreToWig}()}
}
\concept{hetScore}
