% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkSegmentation}
\alias{checkSegmentation}
\title{check segmentation data}
\usage{
checkSegmentation(segmentation)
}
\arguments{
\item{segmentation}{genome segmented into predicted intervals of equal read depth}
}
\value{
data.frame with only those required and recommended columns, others are removed
}
\description{
helper function to check for required columns: 'chr', 'start', 'end', 'rd' and recommend column: 'cnvState'
\itemize{
\item chr, start, end provide the location of the segment
\item rd is the read depth of the segment
\item cnvState an integer 1=loss, 2=normal, 3=gain
}
}
