% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{commonParameters}
\alias{commonParameters}
\title{A dummy function holding a list of all common parameters so
the documentation can be used from one place.}
\usage{
commonParameters(
  sampleId,
  alternateId,
  segmentation,
  segmentationBinSize,
  readDepthPer30kbBin,
  readDepthPer100kbBin,
  centroArray,
  hetScoreData,
  numChroms,
  dPeaksCutoff,
  penaltyCoefForAddingGrids,
  minGridHeight,
  grabDataPercentManual,
  origMaxPercentCutoffManual,
  pause,
  noPdf,
  skipExtras,
  minPeriodManual,
  maxPeriodManual,
  forceFirstDigPeakCopyNum,
  minReasonableSegmentSize,
  outputDir,
  heterozygosityScoreThreshold,
  hsNormMat
)
}
\arguments{
\item{sampleId}{sample Identifier}

\item{alternateId}{secondary sample identifier}

\item{segmentation}{identified regions of the genome with constant read depth. Contains chromosome, start, end, expected CNV, actual CNV and other values we do not need.}

\item{segmentationBinSize}{bin size used for the read depth in the segmentation data}

\item{readDepthPer30kbBin}{read depth for 30 kb bins, with bin index in linear coordinates}

\item{centroArray}{array with the positions of the centromeres for each chromosome}

\item{hetScoreData}{heterozygosity scores determined per 30 kb bin over a 1 Mb region}

\item{numChroms}{number of chromosomes in the reference genome to consider}

\item{dPeaksCutoff}{dPeaksCutoff min grid height for a peak to be considered a digital peak}

\item{penaltyCoefForAddingGrids}{penalty for adding additional peaks to the digital peak alignment}

\item{minGridHeight}{minimum value that can be assigned to the gridHeights}

\item{grabDataPercentManual}{portion of main peak data to grab, other peaks will be scaled based on read depth (x location), set to -1 to base off of mainPeak width}

\item{origMaxPercentCutoffManual}{peaks smaller than this portion of the max peak are not considered; set to -1 to use default value}

\item{pause}{pause execution until user prompts to continue, available interactively only, useful during testing}

\item{skipExtras}{logical to turn on/off plots used for testing and debugging}

\item{minPeriodManual}{manually set \code{minPeriod} within \code{calculatePloidy}}

\item{maxPeriodManual}{manually set \code{maxPeriod} within \code{calculatePloidy}}

\item{forceFirstDigPeakCopyNum}{value to force copy number of first digital peak, use only when ploidy calculation is wrong}

\item{minReasonableSegmentSize}{initial smallest segment size to include in ploidy test segments; want to keep as large as possible to avoid 0N segments, but will decrease size if not enough segments are found}

\item{hsNormMat}{}

\item{ideogram}{The \link{ideogram} object. If not present, we attempt to use referenceGenomeDescriptor}
}
\description{
A dummy function holding a list of all common parameters so
the documentation can be used from one place.
}
