% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{convertChromToCharacter}
\alias{convertChromToCharacter}
\title{Convert chromosome integer to a character string}
\usage{
convertChromToCharacter(ichr, rgdObject = NULL, withChrPrefix = FALSE)
}
\arguments{
\item{ichr}{number vector, for example:  2 or c(2,5,23)}

\item{rgdObject}{Reference genome descriptor, if NULL assumes autosomes= chr1-22 and allosomes= chr23,24}

\item{withChrPrefix}{if TRUE, will add "chr" to the front of every ichr value}
}
\value{
character vector of ichr
}
\description{
allosome 23 will be reported as X, 24 as Y, 25 as M.
Option to include "chr" prefix, if FALSE and if the 'chr' prefix is present, 'chr' will be removed.
}
