% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{digitalGrid}
\alias{digitalGrid}
\title{fit a group of peaks to a digital grid}
\usage{
digitalGrid(
  peakInfo,
  gridHeights,
  dPeaksCutoff,
  penaltyCoefForAddingGrids,
  n00,
  someNumber = 8,
  sampleId = NULL,
  alternateId = NULL,
  gridIteration,
  minPeriodManual = minPeriodManual,
  maxPeriodManual = maxPeriodManual,
  numOfGridCoordsToTest = NULL,
  previousPeriod = NULL,
  minGridHeight,
  iterationStats,
  omitAnnotations = FALSE
)
}
\arguments{
\item{someNumber}{}

\item{minPeriodManual}{user provided minPeriod, default is -1 to indicate no user input}

\item{absMinPeriod}{the smallest minPeriod allowed}
}
\value{
dPeaks-the coordinates of the digital peaks, where NA is a peak not on the digital grid, and nCopyPeaks_dig-first digital peak set to 1
}
\description{
fit a group of peaks to a digital grid
}
