% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{extrapRD}
\alias{extrapRD}
\title{extrapolate or interpolate read depth for a give copy number}
\usage{
extrapRD(cn, cn1, cn2, rd1, rd2)
}
\description{
given read depth and copy number for two other positions
}
\examples{
extrapRD(cn=2,cn1=4,cn2=5,rd1=2,rd2=2.376)

}
\keyword{internal}
