% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{getCentromerePositions}
\alias{getCentromerePositions}
\title{Get centromere positions, will return NULL if no centromeres exist (i.e. in mouse, rat)}
\usage{
getCentromerePositions()
}
\value{
Centromere position array. Two-dimensions - first is chromosome number, the other separates start and end positions.
}
\description{
Get centromere positions, will return NULL if no centromeres exist (i.e. in mouse, rat)
}
\concept{ideogram}
