% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getCNcolors}
\alias{getCNcolors}
\title{list of colors to use for the peaks in \code{peaksByDensity}}
\usage{
getCNcolors()
}
\description{
list of colors to use for the peaks in \code{peaksByDensity}
}
