% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getDiploidPeakNRD}
\alias{getDiploidPeakNRD}
\title{NRD of the diploid peak from the binned read depth data}
\usage{
getDiploidPeakNRD(result)
}
\description{
NRD of the diploid peak from the binned read depth data
}
