% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getLohContent}
\alias{getLohContent}
\title{get LOH content metrics used to identify high-ploidy}
\usage{
getLohContent(allelicSegData)
}
\arguments{
\item{allelicSegData}{segment data returned from \code{calculatePloidy} and augmented with major and minor allele specific copy number}
}
\value{
a list with the three metrics used by us (A) and other authors (B and C)
lohContentA_maj2_min0
lohContentB_maj1_min0
lohContentC_maj2
}
\description{
get LOH content metrics used to identify high-ploidy
}
