% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getMainPeakNRD}
\alias{getMainPeakNRD}
\title{NRD of the main Peak in the ---cnvBinned data---
Will equal 2 if main peak is the normal peak}
\usage{
getMainPeakNRD(result)
}
\arguments{
\item{result}{output from \code{calculatePloidy()}}
}
\description{
NRD of the main Peak in the ---cnvBinned data---
Will equal 2 if main peak is the normal peak
}
