% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getModeOrMaxScore}
\alias{getModeOrMaxScore}
\title{Calculate density of loh scores}
\usage{
getModeOrMaxScore(
  dataIn,
  considerPeakCutoff = 0.06,
  countPeakCutoff = 0.25,
  plotTextPrefix = NULL,
  folderId = NULL,
  sampleId = NULL,
  skipPlot = FALSE,
  minObserv = 15,
  default = 0,
  heterozygosityScoreThreshold
)
}
\arguments{
\item{considerPeakCutoff}{peaks must be bigger than this percentage of maxPeak to be considered # TODO: should this be adjusted higher when there are more data points, ie >100 or 125? 26297 requires >=0.06 not 0.05 which is what I've done all my testing on}

\item{countPeakCutoff}{peaks must be bigger than this percentage of maxPeak to be counted for numClusters}

\item{plotTextPrefix}{text to add to front of plot title}

\item{skipPlot}{logical to control making density plot}

\item{minObserv}{}

\item{default}{default return value}

\item{data}{lohScores mean het scores for selected copy number segments}
}
\description{
If enough data ( >minObserv ) return testScore=right-most mode that is big enough (> consider PeakCutoff * max peak) and numHetClusters > 20\% max peak
If not enough data return max loh score and numHetClusters =NA
}
