% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getPloidyExistsStatus}
\alias{getPloidyExistsStatus}
\title{determine if ploidy was run or not}
\usage{
getPloidyExistsStatus(postProcessingDir_SV, sampleId_SV)
}
\value{
TRUE if ploidy was run FALSE if ploidy was not run
}
\description{
determine if ploidy was run or not
}
