% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getPloidyParamsForSample}
\alias{getPloidyParamsForSample}
\title{retrieve recommended non-default config inputs to calculatePloidy() for a given folder}
\usage{
getPloidyParamsForSample(numfolder)
}
\arguments{
\item{numfolder}{5-digit folder number for a sample as provided in samplenames.csv loaded by \code{getSampleDatabase()}}
}
\value{
dataframe with param and value, row name will be the folder
}
\description{
retrieve recommended non-default config inputs to calculatePloidy() for a given folder
}
