% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyState.R
\name{getPloidyState}
\alias{getPloidyState}
\title{Determine ploidy state from ploidy and 2N+LOH content}
\usage{
getPloidyState(twoNpLOH, ploidy)
}
\arguments{
\item{twoNpLOH}{2N+LOH content determined from \code{plotStarsInTheClouds}}

\item{ploidy}{ploidy value determined from \code{plotStarsInTheClouds}}
}
\value{
a vector the same length as the input, with the value 'near-diploid' or 'high-ploidy'
}
\description{
based on SVM line separating the two clusters n=603
y = -0.411x + 1.126  # Johnson/Arnav 2023, n=603
}
\examples{
getPloidyState(twoNpLOH=.2, ploidy=2.5)
getPloidyState(twoNpLOH=.03, ploidy=2.6)

}
