% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSegmentColors}
\alias{getSegmentColors}
\title{chromosome segment colors
one color for each chromosome 1-22}
\usage{
getSegmentColors()
}
\description{
chromosome segment colors
one color for each chromosome 1-22
}
