% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getTwoBestPeakIndexes}
\alias{getTwoBestPeakIndexes}
\title{find the two best digital peaks for calculating tumor percentage or extrapolating other peaks}
\usage{
getTwoBestPeakIndexes(peakCopyNum, peakHeight)
}
\arguments{
\item{peakCopyNum}{from peakInfo, copy number for the peak}

\item{peakHeight}{from peakInfo, height of each peak normalized so max peak = 1}
}
\description{
Best peak will be the 2N peak if it exists and is suitable, otherwise use main peak
Second best peak will be the taller of the two digital peaks adjacent to the best peak
Don't rely on 2N and then the 1N/3N because they may be small and/or missing
2N peak is considered suitable if it is one of the top 3 biggest peaks
Do not default to two biggest peaks, we want to use lower copy number peaks which will typically align to the digital grid better due to less (sub) clonal mixing
}
