% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGenomePlot.R
\name{grayColorVector}
\alias{grayColorVector}
\title{Provide gray color coding only for display on the genome plot}
\usage{
grayColorVector(x)
}
\arguments{
\item{x}{linear genome x-axis, windows, from the read-depth binned data}
}
\description{
Provide gray color coding only for display on the genome plot
}
