% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGenomePlot.R
\name{linearGenomePlot}
\alias{linearGenomePlot}
\title{Draw a linear genome plot}
\usage{
linearGenomePlot(
  readDepthBinnedData,
  readDepthBinSize = 30000,
  yLimQuantile = 0.99,
  sampleId = NULL,
  segmentation = NULL,
  noDelAmpDetection = FALSE,
  gainColor = "blue",
  lossColor = "red",
  ...
)
}
\arguments{
\item{readDepthBinnedData}{list of two arrays: one with (normalized) read count per each window, second with linear genome position of each window (masked windows have been removed)}

\item{readDepthBinSize}{size of each window}

\item{yLimQuantile}{To prevent outliers from dominating the plot, we set the y axis to go up to
this quantile (number 0-1, 1=100\\%, 0.5=median)}

\item{sampleId}{plot y axis with normalize read depth to 2 for the median frequency}

\item{segmentation}{read depth data.frame with required columns: chr, start, end, rd; optional: cnvState for color coded linear genome plot}

\item{noDelAmpDetection}{do not color code deletions and gains in genome plot}

\item{gainColor}{color to use for gains in linear genome plot, default is blue}

\item{lossColor}{color to use for losses in linear geneome plot, default is red}

\item{...}{Parameters passed onto the actual plot command}
}
\description{
Displays the read-depth array in a linear fashion. Can report CNV as all black or use color calls.
If sampleId is provided, title annotations will be added. But should be specified in order to retrieve other data (segmentation, allelicSegData, normalPeakMethod),
unless those are passed in too.
Does not display structural variants.
}
\examples{
sampleId='TCGA-14-1402-02A_ds'
inputDir='/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Rprojects/BACDAC/inst/extdata'
outputDir='/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Routput/BACDAC'
segmentationFile = file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
segmentation <- read.csv(segmentationFile,comment.char = '#', header = TRUE); dim(segmentation)
thirtyKbFile=file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds'))
readDepthBinnedData = readRDS(file=thirtyKbFile )

 op <- par(mfrow=c(3,1),mai=c(.25,0.5, 0.3,0.25), mgp=c(2, .5, 0))
# default cnv coloring (by default) and annotations
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=sampleId,segmentation=segmentation)
# example with no cnv coloring
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=sampleId,noDelAmpDetection=TRUE,segmentation=segmentation)
# reviewer preferred cnv coloring
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=NULL,segmentation=segmentation, gainColor = 'red', lossColor= 'blue' )

}
