% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{linearToBima}
\alias{linearToBima}
\title{Translate linear genome coordinates back to reference genome coordinates}
\usage{
linearToBima(rgd, globalPos, binSize = 1)
}
\arguments{
\item{rgd}{Reference Genome Descriptor}

\item{globalPos}{linear genome position (1-based)}

\item{binSize}{Size of bin that was originally used

#' @return A list with two elements:
\tabular{ll}{
\code{svaNumber} \tab reference sequence number \cr
\code{svaPos} \tab position within the reference sequence (1-based) \cr
}}
}
\description{
This is an inverse function to \link{bimaToLinear}.
}
\seealso{
Other coordinates: 
\code{\link{bimaToLinear}()},
\code{\link{binnedPosEnd}()},
\code{\link{binnedPosStart}()},
\code{\link{getLinearCoordinates}()},
\code{\link{linearCoordinates}}
}
\concept{coordinates}
