% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadHsNormMat}
\alias{loadHsNormMat}
\title{load hsNormMat}
\usage{
loadHsNormMat(hsNormMatFile)
}
\arguments{
\item{hsNormMatFile}{full path to hsNormMatFile}
}
\description{
if file does not exist, will download from zenodo. Will check contents and return matrix if
successfully loaded.
}
