% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSegmentationFile}
\alias{loadSegmentationFile}
\title{load segmentation file
check for required columns}
\usage{
loadSegmentationFile(segmentationFile)
}
\arguments{
\item{segmentationFile}{full path to file to load}
}
\description{
load segmentation file
check for required columns
}
