% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadTestVals}
\alias{loadTestVals}
\title{load testVals}
\usage{
loadTestVals(testValsFile)
}
\arguments{
\item{testValsFile}{full path to testValsFile}
}
\description{
if file does not exist, will download from zenodo. Will check contents and return matrix if
successfully loaded.
}
