% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{makeAndSaveHetScorePerArm}
\alias{makeAndSaveHetScorePerArm}
\title{summarize the hetScore from bins to chromosome arm}
\usage{
makeAndSaveHetScorePerArm(
  hetScorePerArmFile,
  seqValsTotal,
  chromsToSave = 1:23,
  noPArm = c(13, 14, 15, 21, 22)
)
}
\arguments{
\item{hetScorePerArmFile}{full path to Heterozygosity Score per arm csv file as created by \code{calculateHetScore}}

\item{seqValsTotal}{Calculated values, 30K binned, list of one array per chromosome}

\item{chromsToSave}{List of chromosomes to calculate the value for (by default all but Y)}

\item{noPArm}{Do not return calculation for these p arms}
}
\value{
data.frame with chr, arm, hetScore columns
}
\description{
summarize the hetScore from bins to chromosome arm
}
\seealso{
Other hetScore: 
\code{\link{calculateHetScoreTestStatisticPerCoverage}()},
\code{\link{plotHetScorePerArm}()},
\code{\link{plotHetScorePerBin}()},
\code{\link{saveHetScoreToWig}()}
}
\concept{hetScore}
