% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGenomePlot.R
\name{makeCNVcolorVector}
\alias{makeCNVcolorVector}
\title{Provide the color coding for display on the genome plot}
\usage{
makeCNVcolorVector(x, delAmp = NULL, gainColor = "blue", lossColor = "red")
}
\arguments{
\item{x}{linear genome x-axis, windows, from the read-depth binned data}

\item{delAmp}{segmentation converted to the legacy delAmp format using \code{makeLegacyDelAmp}}

\item{gainColor}{color to use for gains in linear genome plot, default is blue}

\item{lossColor}{color to use for losses in linear geneome plot, default is red}
}
\value{
a vector of color codes to reflect the CNV calls: 2=red/loss, 4=blue/gains, 8=gray/normal; or 1=black/no color coding
}
\description{
Provide the color coding for display on the genome plot
}
\keyword{internal}
