% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{makeHetScoreReportPdf}
\alias{makeHetScoreReportPdf}
\title{Make 3-subplot hetScore analysis PDF}
\usage{
makeHetScoreReportPdf(
  hetScorePerBinWigFile,
  hetScorePerArmFile,
  segmentation = NULL,
  allelicSegData = NULL,
  sampleId,
  outputDir,
  noPdf
)
}
\arguments{
\item{hetScorePerBinWigFile}{full path to Heterozygosity Score per bin wig file as created by \code{calculateHetScore}}

\item{hetScorePerArmFile}{full path to Heterozygosity Score per arm csv file as created by \code{calculateHetScore}}

\item{segmentation}{read depth data.frame with required columns: chr, start, end, rd}

\item{allelicSegData}{allele specific segmentation file}

\item{sampleId}{sample id, will be used as the prefix for all input and output file names}

\item{outputDir}{full path for output files}

\item{noPdf}{When TRUE, pdf files will not be generated, instead plots are drawn on default device}
}
\description{
Will show 3 separate plots contrasting CNV, heterozygosity score per 30K and heterozygosity score per arm.
}
