% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGenomePlot.R
\name{makeLegacyDelAmp}
\alias{makeLegacyDelAmp}
\title{for compatibility with genomePlot. Make George's old array from new structure: 1=normal, -1=loss, 2=gain}
\usage{
makeLegacyDelAmp(segmentation, rgdObject, coords)
}
\arguments{
\item{segmentation}{segmentation loaded via \link{typedFile-class}}

\item{rgdObject}{reference genome descriptor object loaded via \code{loadRgd()}}

\item{coords}{linear genome coordinates, loaded via \code{getLinearCoordinates(rgd=rgdObject)}}
}
\description{
for compatibility with genomePlot. Make George's old array from new structure: 1=normal, -1=loss, 2=gain
}
