% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationUtils.R
\name{makeStarLookUpTable}
\alias{makeStarLookUpTable}
\title{create the starLookUp table}
\usage{
makeStarLookUpTable(starCloudResult, percentTumor)
}
\arguments{
\item{starCloudResult}{object returned by \code{plotStarsInTheClouds}}

\item{percentTumor}{tumor percent listed in the object returned by \code{calculatePloidy}}
}
\value{
A data.frame with 5 columns: hetScore, nrd, cn, major, minor.
One row copy number 1 to max copy
}
\description{
lists the hetScore for each possible allele fraction
}
\keyword{internal}
