% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{markChromEdges}
\alias{markChromEdges}
\title{add vertical lines to linear genome plot}
\usage{
markChromEdges(chromStarts, maxcn, vCol = "gray90")
}
\arguments{
\item{chromStarts}{must be same window size scale as the plotted frequency data}

\item{maxcn}{max chromosome number}
}
\description{
add vertical lines to linear genome plot
}
