% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{peaksByDensity}
\alias{peaksByDensity}
\title{Use density on copy number (read depth) to determine peaks}
\usage{
peaksByDensity(
  sampleId,
  readDepthPer100kbBin,
  segmentation,
  segmentationBinSize = 30000,
  wszPeaks = 1e+05,
  grabDataPercentManual = -1,
  origMaxPercentCutoffManual = -1,
  addAreaLinesToPlot = FALSE,
  qualityPostNorm = NULL,
  pause = FALSE,
  omitAnnotations = FALSE,
  alternateId = NULL
)
}
\arguments{
\item{sampleId}{sample Identifier}

\item{segmentation}{identified regions of the genome with constant read depth}

\item{segmentationBinSize}{bin size used for the read depth in the segmentation data}

\item{wszPeaks}{window size for peaks by density}

\item{grabDataPercentManual}{portion of main peak data to grab, other peaks will be scaled based on read depth (x location) set to -1 to use main Peak width}

\item{origMaxPercentCutoffManual}{peaks smaller than this portion of the max peak are not considered; set to -1 to use default value}

\item{qualityPostNorm}{used to determine grabDataPercentMax}

\item{alternateId}{secondary sample identifier}

\item{rgdObject}{loaded rgd file as an object}

\item{cnvBinnedData}{binned copy number variant (actually read depth) values, linear coordinates, with normal}
}
\description{
peaks are returned in order of read depth with most common (highest frequency) read depth peak listed first
}
