% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{ploidyCheck}
\alias{ploidyCheck}
\title{calculate the error in hetScore between the loh segments and the corresponding theoretical value}
\usage{
ploidyCheck(starCloudResult, segmentData, percentTumor)
}
\arguments{
\item{starCloudResult}{result output from \code{plotStarsInTheClouds}}

\item{segmentData}{segments used in calculatePloidy which has meanLOH per segment}

\item{percentTumor}{as output from calculateploidy}
}
\description{
if the error is bigger than a set negative value, ploidy passes.
Any positive value is good, means the clouds are all right of the loh (purple) line
RMSD is not appropriate for this test because RMSD squares the data to get only positive values,
we want to check for too large of a negative difference, large positive differences may occur especially at high tumor
when the clouds shift right of the loh line but this does not signal an error in the ploidy determination
}
