% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGenomePlot.R
\name{plotEmptyLinearGenomePlot}
\alias{plotEmptyLinearGenomePlot}
\title{create an empty linear genome plot to fill a space when data is not available}
\usage{
plotEmptyLinearGenomePlot(chromsToPlot, coords)
}
\arguments{
\item{chromsToPlot}{vector of chromosome (as integers) to plot}

\item{coords}{object describing linear coordinate space for the chromosomes}
}
\description{
create an empty linear genome plot to fill a space when data is not available
}
\keyword{internal}
