% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyState.R
\name{plotPloidyState}
\alias{plotPloidyState}
\title{plot ploidy vs 2N+LOH}
\usage{
plotPloidyState(twoNpLOH, ploidy, maxPloidy = 6)
}
\arguments{
\item{twoNpLOH}{2N+LOH content determined from \code{plotStarsInTheClouds}}

\item{ploidy}{ploidy value determined from \code{plotStarsInTheClouds}}

\item{maxPloidy}{max x axis value for plot, default=6}
}
\description{
includes ploidy state separation line
}
\examples{
plotPloidyState(twoNpLOH=.2, ploidy=2.5)
plotPloidyState(twoNpLOH=.03, ploidy=2.6)

}
