% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationPlot.R
\name{plotStarsInTheClouds}
\alias{plotStarsInTheClouds}
\title{plot heterozygosity (actual and theoretical) vs NRD for a given tumor ratio}
\usage{
plotStarsInTheClouds(
  sampleId,
  alternateId,
  starCloudPlotInputs,
  diploidPeakNRD,
  tau,
  plotEachChrom = FALSE,
  mainPeakNRD,
  segmentData = NULL,
  peakInfo = NULL,
  forceFirstDigPeakCopyNum = -1,
  grabDataPercentManual = -1,
  origMaxPercentCutoffManual = -1,
  minPeriodManual = -1,
  maxPeriodManual = -1,
  minReasonableSegmentSize = 5500000,
  digitalPeakZone = 0.05,
  heterozygosityScoreThreshold = 0.98,
  paperMode = FALSE
)
}
\arguments{
\item{diploidPeakNRD}{the NRD of the diploid peak in  ---cnvBinned data---, may not be 2 if you choose a different peak from the pipeline to normalize by}

\item{tau}{tumor ratio}

\item{mainPeakNRD}{the NRD of the main Peak in the ---cnvBinned data---, if mainPeak based, will be 2, but if ploidybased, may be different}
}
\description{
plot heterozygosity (actual and theoretical) vs NRD for a given tumor ratio
}
\examples{
sampleId='TCGA-14-1402-02A_ds'; alternateId=66301
tau=min(1,calcPloidyResult$percentTumor/100); segmentData=calcPloidyResult$segmentData; peakInfo=calcPloidyResult$peakInfo
mainPeakNRD= getMainPeakNRD(calcPloidyResult); diploidPeakNRD=NULL; #getDiploidPeakNRD(result)
digitalPeakZone =calcPloidyResult[['iterationStatsAll']][['digitalPeakZone']]
op <- par(mfrow=c(1,1),mar=c(5,4,3.5,3.5),mgp=c(1.5, 0.5,0))
plotStarsInTheClouds(sampleId, alternateId, starCloudPlotInputs, diploidPeakNRD, tau, plotEachChrom=TRUE, mainPeakNRD,segmentData=segmentData, peakInfo=peakInfo)
par(op)
}
