% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBACDAC.R
\name{runBACDAC}
\alias{runBACDAC}
\title{run all the steps in BACDAC, will output pdfs and intermediate files if outputDir is provided}
\usage{
runBACDAC(
  sampleId,
  alternateId,
  outputDir,
  noPdf = FALSE,
  readDepthPer30kbBin,
  readDepthPer100kbBin,
  segmentation,
  segmentationBinSize,
  hsNormMat,
  testVals,
  dPeaksCutoff = 0.01,
  penaltyCoefForAddingGrids = 0.49,
  minGridHeight = 0.2,
  minPeriodManual = -1,
  maxPeriodManual = -1,
  forceFirstDigPeakCopyNum = -1,
  grabDataPercentManual = -1,
  origMaxPercentCutoffManual = -1,
  minReasonableSegmentSize = 5500000,
  heterozygosityScoreThreshold = 0.98,
  allowedTumorPercent = 106
)
}
\arguments{
\item{sampleId}{sample Identifier}

\item{alternateId}{optional secondary sample identifier}

\item{outputDir}{output directory}

\item{noPdf}{if present, do not create pdfs}

\item{readDepthPer30kbBin}{read depth for 30 kb bins, with bin index in linear coordinates}

\item{readDepthPer100kbBin}{read depth for 100 kb bins, with bin index in linear coordinates}

\item{segmentation}{identified regions of the genome with constant read depth. Contains chromosome, start, end, expected CNV, actual CNV and other values we do not need.}

\item{segmentationBinSize}{bin size used for the read depth in the segmentation data}

\item{hsNormMat}{hetScores from  a database of 23 Normals,  101046 x 23 rows, one row for each 30 kb segment of the genome, 1-22, X and a part of Y. Columns are values for each of the 23 Normals for each segment}

\item{dPeaksCutoff}{dPeaksCutoff min grid height for a peak to be considered a digital peak}

\item{penaltyCoefForAddingGrids}{penalty for adding additional peaks to the digital peak alignment}

\item{minGridHeight}{minimum value that can be assigned to the gridHeights}

\item{minPeriodManual}{user provided \code{minPeriod} for digital grid, default is -1 to indicate no user input}

\item{maxPeriodManual}{user provided \code{maxPeriod} for digital grid, default is -1 to indicate no user input}

\item{forceFirstDigPeakCopyNum}{value to force copy number of first digital peak, use only when ploidy calculation is wrong}

\item{grabDataPercentManual}{portion of main peak data to grab, other peaks will be scaled based on read depth (x location), set to -1 to base off of mainPeak width}

\item{origMaxPercentCutoffManual}{peaks smaller than this portion of the max peak are not considered; set to -1 to use default value}

\item{minReasonableSegmentSize}{initial smallest segment size to include in ploidy test segments; want to keep as large as possible to avoid 0N segments, but will decrease size if not enough segments are found}

\item{heterozygosityScoreThreshold}{peaks with a hetScore mode above this value are considered heterozygous, typically 0.98, but may vary depending on NGS library quality and preparation}
}
\description{
run all the steps in BACDAC, will output pdfs and intermediate files if outputDir is provided
}
\examples{
\dontrun{
library(BACDAC)
library(logging)
basicConfig("DEBUG")
sampleId='TCGA-14-1402-02A_ds'; alternateId=66301
outputDir <- file.path('/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Routput/BACDAC', sampleId) # tempdir()
noPdf=TRUE                          # TRUE= print to screen, FALSE=print to pdf (i e. outputDir/dev/ploidy)

# inputDir is the path to the load package data
inputDir <- system.file('extdata', package = "BACDAC")

# read depth data
thirtyKbFile=file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds'))
readDepthPer30kbBin = readRDS(file=thirtyKbFile )
hundredKbFile=file.path(inputDir, paste0(sampleId,'_','readDepthPer100kbBin.Rds'))
readDepthPer100kbBin = readRDS(file=hundredKbFile )

# segmentation data
segmentationFile <- file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
segmentation= loadSegmentationFile(segmentationFile)

## load two reference files  ---------------
# hsNormMat/lohMat: hetScores from 23 Normals,  101046 x 23 rows, one row for each 30kb segment of the genome, 1-22, X and a part of Y. Columns are values for each of the 23 Normals for each segment
#                   used to look for places in 23 TCGA normals where more than half dropped below the a (i.e. 0.975) cutoff.
# testVals: used to find each possible heterozygosity value for each copy number level (find the right spots for the stars)
hsNormMat <- bmdTools::loadRdata('/research/labs/experpath/vasm/shared/NextGen/Misc/pipelineInputs/hetScoreAnalysis/lohMat.Rdata') # aka lohMat
testVals <-  bmdTools::loadRdata(file.path('/research/labs/experpath/vasm/shared/NextGen/Misc/pipelineInputs/hetScoreAnalysis/testVals.Rdata'))

runBACDAC(sampleId, alternateId,
          outputDir,
          noPdf,
          readDepthPer30kbBin, readDepthPer100kbBin,segmentation,hsNormMat,testVals)
}
}
