% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{saveHetScoreToWig}
\alias{saveHetScoreToWig}
\title{Save the binned results of \code{calculateHetScore} as a wig file}
\usage{
saveHetScoreToWig(
  wigFile,
  seqListTotal,
  seqValsTotal,
  chromsToSave,
  samplingStep
)
}
\arguments{
\item{wigFile}{full Path to wig output file containing the Heterozygosity Score per bin as calculated by \code{calculateHetScore}}

\item{seqListTotal}{Start position for each 30K bin, 1 based, list of one array per chromosome}

\item{seqValsTotal}{Calculated values, 30K binned, list of one array per chromosome}

\item{chromsToSave}{List of chromosomes to calculate the value for (by default all but Y)}

\item{samplingStep}{Only uniformly sampled data can be used, use this sampling step}
}
\description{
Save the binned results of \code{calculateHetScore} as a wig file
}
\seealso{
Other hetScore: 
\code{\link{calculateHetScoreTestStatisticPerCoverage}()},
\code{\link{makeAndSaveHetScorePerArm}()},
\code{\link{plotHetScorePerArm}()},
\code{\link{plotHetScorePerBin}()}
}
\concept{hetScore}
\keyword{internal}
