% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{seqFwd}
\alias{seqFwd}
\title{Forward only \code{:} replacement}
\usage{
seqFwd(from = 1, to = 1)
}
\arguments{
\item{from}{first sequence item}

\item{to}{last sequence item}
}
\value{
Just like \code{from:to} except if \code{to < from}, the returned
sequence is empty.
}
\description{
Forward only \code{:} replacement
}
