% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationPlot.R
\name{twoPanelReport}
\alias{twoPanelReport}
\title{draw constellation plot left of the linear genome plot}
\usage{
twoPanelReport(
  starCloudPlotInputs,
  calcPloidyResult,
  readDepthPer30kbBin,
  segmentation,
  sampleId = NULL,
  alternateId = NULL,
  diploidPeakNRD = NULL,
  gainColor = "blue",
  lossColor = "red"
)
}
\description{
draw constellation plot left of the linear genome plot
}
