% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{allelicCNV}
\alias{allelicCNV}
\title{allele specific cnv
\enumerate{
\item find CN for the segment by rounding the cnLevel to an integer
\item find which closest star, within that CN, is closest to the hetScore of the segment
\item the closest star is then the segment's major:minor allele
}}
\usage{
allelicCNV(starLookUp, segmentDataIn)
}
\description{
allele specific cnv
\enumerate{
\item find CN for the segment by rounding the cnLevel to an integer
\item find which closest star, within that CN, is closest to the hetScore of the segment
\item the closest star is then the segment's major:minor allele
}
}
