% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{bimaToLinear}
\alias{bimaToLinear}
\title{Translate from BIMA coordinates to linear genome.}
\usage{
bimaToLinear(svaNumber, svaPos, binSize = 1)
}
\arguments{
\item{svaNumber}{BIMA chromosome number (as present in the SVA file)}

\item{svaPos}{Position within the sequence (1-based)}

\item{binSize}{How to round the coordinates. The result will be divided by this number.
Note - recreates a bug in George's code at the moment,
so the result can be up to binSize away from what you'd expect.}
}
\value{
Linear genome position (1-based). For negative \code{svaPos}
return negative number.
}
\description{
Translate a list of BIMA chromosomes (svaNumbers) and BIMA positions to the
linear genome coordinates. The linear genome starts at position 1 and
consists of all sequences in linear order.
}
\seealso{
Other coordinates: 
\code{\link{binnedPosEnd}()},
\code{\link{binnedPosStart}()},
\code{\link{getLinearCoordinates}()},
\code{\link{linearCoordinates}},
\code{\link{linearToBima}()}
}
\concept{coordinates}
