% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationPlot.R
\name{blankPlotStarsInTheClouds}
\alias{blankPlotStarsInTheClouds}
\title{make a blank plot, no data, used as a stand in when combining lots of samples into one pdf}
\usage{
blankPlotStarsInTheClouds(heterozygosityScoreThreshold = 0.98, tau = 0.8)
}
\description{
make a blank plot, no data, used as a stand in when combining lots of samples into one pdf
}
