% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{calcCopyNumber}
\alias{calcCopyNumber}
\title{find copy number from read depth and tumor fraction}
\usage{
calcCopyNumber(NRD, tau)
}
\arguments{
\item{NRD}{normalized read depth}

\item{tau}{tumor fraction}
}
\description{
find copy number from read depth and tumor fraction
}
\examples{
calcCopyNumber(NRD=23,tau=0.85)
calcCopyNumber(NRD=1.4,tau=0.33)
calcCopyNumber(NRD=4,tau=0.75)
}
\seealso{
Other copyNumberCalcs: 
\code{\link{calcNrd}()},
\code{\link{calcTau}()},
\code{\link{calcTumorRatio}()}
}
\concept{copyNumberCalcs}
