% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{calcNrd}
\alias{calcNrd}
\title{find normalized read depth from copy number and tumor ratio}
\usage{
calcNrd(cn, tau)
}
\arguments{
\item{cn}{copy number}

\item{tau}{tumor fraction}
}
\description{
find normalized read depth from copy number and tumor ratio
}
\examples{
calcNrd(cn=3,tau=.7)
}
\seealso{
Other copyNumberCalcs: 
\code{\link{calcCopyNumber}()},
\code{\link{calcTau}()},
\code{\link{calcTumorRatio}()}
}
\concept{copyNumberCalcs}
