% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{calcRD}
\alias{calcRD}
\title{find read depth per window size from NRD}
\usage{
calcRD(nrd, wsz, expReadsIn2NPeak_1bp)
}
\arguments{
\item{nrd}{normalized read depth}

\item{wsz}{window size}

\item{expReadsIn2NPeak_1bp}{expected number of reads in a 1 bp bin for the diploid peak}
}
\description{
find read depth per window size from NRD
}
