% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{calculateHetScore}
\alias{calculateHetScore}
\title{Combines ref/alt SNP data from each chromosome 1-22,X}
\usage{
calculateHetScore(
  sampleId,
  alternateId = NULL,
  inputDir,
  outputDir,
  segmentation = NULL,
  noPdf = FALSE,
  maximumCoverage = 1000,
  trimFromAlt = 2,
  trimFromRef = 1,
  trimExtraPerCoverage = 0.1,
  minSnpsToCalculateStatistic = 20,
  samplingStep = 30000,
  extraWindow = 1e+06,
  readDepthBinnedData = NULL,
  readDepthBinSize = 30000,
  gainColor = "blue",
  lossColor = "red",
  noDelAmpDetection = FALSE
)
}
\arguments{
\item{sampleId}{sample identifier, used as prefix for all input and output file names \code{<sampleId>_snpVals_<chromosome>.Rdata} and \code{<sampleId>_countBP_<chromosome>.Rdata}}

\item{inputDir}{full path to snpVal and countBP Rdata}

\item{outputDir}{full path for output files}

\item{segmentation}{read depth data.frame with required columns: chr, start, end, rd; optional: cnvState for color coded linear genome plot}

\item{noPdf}{When TRUE, pdf files will not be generated, instead plots are drawn on default device}

\item{maximumCoverage}{Do not process SNPs covered more than this}

\item{trimFromAlt}{Bins to discard from the 'alt' side of the distribution}

\item{trimFromRef}{Bins to discard from the 'ref' side of the distribution}

\item{trimExtraPerCoverage}{Fraction of bins to trim per each extra coverage}

\item{minSnpsToCalculateStatistic}{Minimum SNPs required in the window (1Mb) to calculate the statistic}

\item{samplingStep}{How frequently to try to summarize the data (bp), to produce overlapping windows}

\item{extraWindow}{Size of the window, how many bps to look at during each sampling step}

\item{readDepthBinnedData}{list of two arrays: one with (normalized) read count per each window, second with linear genome position of each window (masked windows have been removed)}

\item{readDepthBinSize}{size of each window}

\item{gainColor}{color to use for gains in linear genome plot, default is blue}

\item{lossColor}{color to use for losses in linear geneome plot, default is red}

\item{noDelAmpDetection}{do not color code deletions and gains in genome plot}
}
\description{
Outputs heterozygosity score by bin (.wig) and by arm (.csv) for each chromosome
To plot see \link{plotHetScorePerBin} and the three graph plot function \link{makeHetScoreReportPdf}
}
\examples{
sampleId='TCGA-14-1402-02A_ds'
# inputDir is the path to the load package data
inputDir <- system.file('extdata', package = "BACDAC")
outputDir = tempDir() # outputDir='/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Routput/BACDAC'
segmentationFile <- file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
segmentation= loadSegmentationFile(segmentationFile) # per segment: chr, start, end, rd
# defaults:
noPdf = FALSE; maximumCoverage = 1000; trimFromAlt = 2; trimFromRef = 1; trimExtraPerCoverage = 0.1; minSnpsToCalculateStatistic = 20; samplingStep = 30000; extraWindow = 1000000
readDepthBinnedData=NULL; readDepthBinSize=30000; gainColor='blue'; lossColor='red'; noDelAmpDetection=FALSE

\dontrun{

  library(BACDAC)
  # calculateHetScoreExample.R
  # generate heterozygosity score by bin and by arm
  # make a 3-panel plot, top panel the segmentation data and the next two panels are the resulting hetscore by bin and by arm

  basicConfig("DEBUG")
  sampleId='TCGA-14-1402-02A_ds'; alternateId=66301
  outputDir <- tempdir()

  # inputDir is the path to the load package data
  inputDir <- system.file('extdata', package = "BACDAC")
  segmentationFile <- file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
  segmentation= loadSegmentationFile(segmentationFile) # chr, start, end, rd per segment

  calculateHetScore(
    sampleId=sampleId,
    inputDir=inputDir,
    outputDir=outputDir,
    segmentation=segmentation,
    noPdf = TRUE)

}
}
