% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{calculateHetScore}
\alias{calculateHetScore}
\title{Combines ref/alt SNP data from each chromosome 1-22,X}
\usage{
calculateHetScore(
  sampleId,
  inputDir,
  outputDir = inputDir,
  segmentationFile,
  noPdf = FALSE,
  maximumCoverage = 1000,
  trimFromAlt = 2,
  trimFromRef = 1,
  trimExtraPerCoverage = 0.1,
  minSnpsToCalculateStatistic = 20,
  samplingStep = 30000,
  extraWindow = 1e+06
)
}
\arguments{
\item{sampleId}{sample identifier, used as prefix for all input and output file names \code{<sampleId>_snpVals_<chromosome>.Rdata} and \code{<sampleId>_countBP_<chromosome>.Rdata}}

\item{inputDir}{full path to snpVal and countBP Rdata}

\item{outputDir}{full path for output files}

\item{segmentationFile}{segmentation file for read depth, cnv data with required columns: chr, start, end, rd}

\item{noPdf}{When TRUE, pdf files will not be generated, instead plots are drawn on default device}

\item{maximumCoverage}{Do not process SNPs covered more than this}

\item{trimFromAlt}{Bins to discard from the 'alt' side of the distribution}

\item{trimFromRef}{Bins to discard from the 'ref' side of the distribution}

\item{trimExtraPerCoverage}{Fraction of bins to trim per each extra coverage}

\item{minSnpsToCalculateStatistic}{Minimum SNPs required in the window (1Mb) to calculate the statistic}

\item{samplingStep}{How frequently to try to summarize the data (bp), to produce overlapping windows}

\item{extraWindow}{Size of the window, how many bps to look at during each sampling step}
}
\description{
Outputs heterozygosity score by bin (.wig) and by arm (.csv) for each chromosome
To plot see \link{plotHetScorePerBin} and the three graph plot function \link{makeHetScoreReportPdf}
}
\examples{
sampleId='TCGA-14-1402-02A_ds'
inputDir='/research/labs/experpath/vasm/shared/NextGen/Projects/MethodDev/MD66301/GRCh38/svar-1/loh'
outputDir='/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Routput/BACDAC'
segmentationFile=
}
