% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{commonParameters}
\alias{commonParameters}
\title{A dummy function holding a list of all common parameters so
the documentation can be used from one place.}
\usage{
commonParameters(
  sampleId,
  alternateId,
  inputDir,
  outputDir,
  segmentation,
  segmentationBinSize,
  readDepthPer30kbBin,
  readDepthPer100kbBin,
  readDepthBinSize,
  hetScoreData,
  dPeaksCutoff,
  penaltyCoefForAddingGrids,
  minGridHeight,
  grabDataPercentManual,
  origMaxPercentCutoffManual,
  noPdf,
  minPeriodManual,
  maxPeriodManual,
  forceFirstDigPeakCopyNum,
  minReasonableSegmentSize,
  heterozygosityScoreThreshold,
  hsNormMat,
  segmentData,
  allelicSegments,
  peakInfo,
  n00,
  ideogram,
  lossColor,
  gainColor,
  allowedTumorPercent
)
}
\arguments{
\item{sampleId}{sample Identifier}

\item{alternateId}{optional secondary sample identifier}

\item{inputDir}{full path to directory with input files}

\item{outputDir}{full path to directory to output all files and intermediate objects created}

\item{segmentation}{identified regions of the genome with constant read depth. Data.frame with required columns:
chr, start, end, rd; optional: cnvState (1=loss, 2=normal, 3=gain) for color coded linear linear genome plot}

\item{segmentationBinSize}{bin size used for the read depth in the segmentation data}

\item{readDepthPer30kbBin}{list with two equal length arrays:
\itemize{
\item \code{readDepthArray} read depth for 30 kb bins, normalized for GC content and other artifacts
\item \code{goodWindowArray} linear genome position of each window (bin), masked windows have been removed
}}

\item{readDepthPer100kbBin}{list with two equal length arrays:
\itemize{
\item \code{readDepthArray} read depth for 100 kb bins, normalized for GC content and other sequencing artifacts
\item \code{goodWindowArray} linear genome position of each window (bin) (masked windows have been removed)
}}

\item{readDepthBinSize}{bp size of the bins in the read depth array, expecting 30000 or 100000}

\item{hetScoreData}{heterozygosity scores determined per 30 kb bin over a 1 Mb region}

\item{dPeaksCutoff}{dPeaksCutoff min grid height for a peak to be considered a digital peak}

\item{penaltyCoefForAddingGrids}{penalty for adding additional peaks to the digital peak alignment}

\item{minGridHeight}{minimum value that can be assigned to the gridHeights}

\item{grabDataPercentManual}{portion of main peak data to grab, other peaks will be scaled based on read depth (x location),
default is -1, and value will be base on width of the main (dominate) read depth peak}

\item{origMaxPercentCutoffManual}{peaks smaller than this portion of the max peak are not considered; set to -1 to use default value}

\item{noPdf}{if present (TRUE), pdf files will not be generated, instead plots are drawn on default device}

\item{minPeriodManual}{user provided \code{minPeriod} for digital grid, default is -1 to indicate no user input}

\item{maxPeriodManual}{user provided \code{maxPeriod} for digital grid, default is -1 to indicate no user input}

\item{forceFirstDigPeakCopyNum}{value to force copy number of first digital peak, use only when ploidy calculation is wrong}

\item{minReasonableSegmentSize}{initial smallest segment size to include in ploidy test segments (segmentData); want to
keep as large as possible to avoid 0N segments, but will decrease size if not enough segments are found}

\item{heterozygosityScoreThreshold}{peaks with a hetScore mode above this value are considered heterozygous, typically 0.98,
but may vary depending on NGS library quality and preparation}

\item{hsNormMat}{heterozygosity score mask, used to find genomic positions where the heterozgosity score is atypically low.
\itemize{
\item rows: hetScores for each 30 kb segment of the genome, 1-22, X and a part of Y.
\item columns: one column for each of the 23 normal samples
}}

\item{segmentData}{segments from segmentation that are longer than minReasonableSegmentSize and then broken up into segments no smaller than 3 Mbs}

\item{allelicSegments}{segmentData from \code{calculatePloidy} then augmented in \code{plotStarsInTheClouds} with major_copy_number and minor_copy_number}

\item{peakInfo}{summary table of info for each peak output by \code{calculatePloidy}}

\item{n00}{Precision for creating initial grid}

\item{ideogram}{object loaded with the package, contains genomic reference positions for GRCh38 cytobands}

\item{lossColor}{color for losses in the linear genome plot, default is red}

\item{gainColor}{color for gains in the linear genome plot, default is blue}

\item{allowedTumorPercent}{allow some tolerance to go over 100 especially for PDXs which will very near 100}

\item{testVals}{pre-built array to assist finding each possible heterozygosity value for each copy number level}
}
\description{
A dummy function holding a list of all common parameters so
the documentation can be used from one place.
}
