% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{digitalGrid}
\alias{digitalGrid}
\title{fit a group of peaks to a digital grid}
\usage{
digitalGrid(
  peakInfo,
  gridHeights,
  dPeaksCutoff,
  penaltyCoefForAddingGrids,
  n00,
  someNumber = 8,
  sampleId = NULL,
  alternateId = NULL,
  gridIteration,
  minPeriodManual = minPeriodManual,
  maxPeriodManual = maxPeriodManual,
  numOfGridCoordsToTest = NULL,
  previousPeriod = NULL,
  minGridHeight,
  iterationStats,
  omitAnnotations = FALSE
)
}
\arguments{
\item{peakInfo}{summary table of info for each peak found in \code{peaksByDensity}}

\item{dPeaksCutoff}{min grid height for a peak to be considered a digital peak}

\item{sampleId}{sample Identifier}

\item{alternateId}{secondary sample identifier}

\item{minPeriodManual}{user provided minPeriod, default is -1 to indicate no user input}

\item{absMinPeriod}{the smallest minPeriod allowed}
}
\value{
dPeaks-the coordinates of the digital peaks, where NA is a peak not on the digital grid, and nCopyPeaks_dig-first digital peak set to 1
decision tree for maxPeriodToTry, do not allow period to get bigger than this, set to firstToMainPeakGap when: 5 <= mainPeakIndex < 7
decision tree for minPeriod
}
\description{
fit a group of peaks to a digital grid
}
