% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getDiploidPeakNRD}
\alias{getDiploidPeakNRD}
\title{Find the NRD of the diploid peak from the binned read depth data}
\usage{
getDiploidPeakNRD(calcPloidyResult)
}
\arguments{
\item{calcPloidyResult}{object returned from \code{calculatePloidy}}
}
\description{
Find the NRD of the diploid peak from the binned read depth data
}
