% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{getLinearCoordinates}
\alias{getLinearCoordinates}
\title{Get object describing linear coordinate space for main chromosomes}
\usage{
getLinearCoordinates(rgdObject, chromosomes = 1:24)
}
\arguments{
\item{rgdObject}{sysdata object GRCh38 reference genome descriptor used for translating some chromosome names to numbers}

\item{chromosomes}{List of main chromosomes (technically .sva files) to form the coordinate system from}
}
\value{
\link{linearCoordinates}
}
\description{
Get object describing linear coordinate space for main chromosomes
}
\seealso{
Other coordinates: 
\code{\link{bimaToLinear}()},
\code{\link{binnedPosEnd}()},
\code{\link{binnedPosStart}()},
\code{\link{linearCoordinates}},
\code{\link{linearToBima}()}
}
\concept{coordinates}
