% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getMainPeakNRD}
\alias{getMainPeakNRD}
\title{Find the NRD of the main Peak in the ---cnvBinned data---
Will equal 2 if main peak is the normal peak}
\usage{
getMainPeakNRD(calcPloidyResult)
}
\arguments{
\item{calcPloidyResult}{object returned from \code{calculatePloidy}}
}
\description{
Find the NRD of the main Peak in the ---cnvBinned data---
Will equal 2 if main peak is the normal peak
}
\keyword{internal}
