% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{getModeOrMaxScore}
\alias{getModeOrMaxScore}
\title{Calculate density of heterozygosity scores}
\usage{
getModeOrMaxScore(
  dataIn,
  considerPeakCutoff = 0.06,
  countPeakCutoff = 0.25,
  plotTextPrefix = NULL,
  folderId = NULL,
  sampleId = NULL,
  skipPlot = FALSE,
  minObserv = 15,
  default = 0,
  heterozygosityScoreThreshold
)
}
\arguments{
\item{dataIn}{hetScores mean hetScores for selected copy number segments}

\item{considerPeakCutoff}{peaks must be bigger than this percentage of maxPeak to be considered}

\item{countPeakCutoff}{peaks must be bigger than this percentage of maxPeak to be counted for numClusters}

\item{plotTextPrefix}{text to add to front of plot title}

\item{folderId}{alternate id}

\item{sampleId}{sample Identifier}

\item{skipPlot}{logical to control making density plot}

\item{minObserv}{minimum number of hetScores to compute the mode of a peak, otherwise the max value is used}

\item{default}{default return value}

\item{heterozygosityScoreThreshold}{peaks with a hetScore mode above this value are considered heterozygous, typically 0.98,
but may vary depending on NGS library quality and preparation}
}
\description{
If enough data ( >minObserv ) return testScore=right-most mode that is big enough (> consider PeakCutoff * max peak) and numHetClusters > 20\% max peak
If not enough data return max heterozygosity score and numHetClusters =NA
}
\keyword{internal}
