% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdUtils.R
\name{linearCoordinates}
\alias{linearCoordinates}
\title{Linear coordinates class constructor}
\usage{
linearCoordinates(chroms, chromStart, chromEnd)
}
\arguments{
\item{chroms}{A vector of reference sequence numbers for the chromosomes being used}

\item{chromStart}{A vector of all chromosome start positions in linear coordinates.}

\item{chromEnd}{A vector of all chromosome positions in linear coordinates}
}
\value{
\link{linearCoordinates} class
}
\description{
Linear coordinates class constructor
}
\seealso{
Other coordinates: 
\code{\link{bimaToLinear}()},
\code{\link{binnedPosEnd}()},
\code{\link{binnedPosStart}()},
\code{\link{getLinearCoordinates}()}
}
\concept{coordinates}
