% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGenomePlot.R
\name{linearGenomePlot}
\alias{linearGenomePlot}
\title{Draw a linear genome plot}
\usage{
linearGenomePlot(
  readDepthBinnedData,
  readDepthBinSize = 30000,
  sampleId = NULL,
  alternateId = NULL,
  segmentation = NULL,
  allelicSegments = NULL,
  noDelAmpDetection = FALSE,
  gainColor = "blue",
  lossColor = "red",
  zebraStrips = FALSE,
  yAxisLimits = NULL,
  ...
)
}
\arguments{
\item{readDepthBinnedData}{list of two arrays: one with (normalized) read count per each window, second with linear genome position of each window (masked windows have been removed)}

\item{readDepthBinSize}{size of each window}

\item{sampleId}{sample Identifier}

\item{alternateId}{optional secondary sample identifier}

\item{segmentation}{read depth data.frame with required columns: chr, start, end, rd; optional: cnvState for color coded linear genome plot}

\item{allelicSegments}{segmentData from \code{calculatePloidy} then augmented in \code{plotStarsInTheClouds} with major_copy_number and minor_copy_number}

\item{noDelAmpDetection}{do not color code deletions and gains in genome plot}

\item{gainColor}{color to use for gains in linear genome plot, default is blue}

\item{lossColor}{color to use for losses in linear genome plot, default is red}

\item{zebraStrips}{option to have alternating gray/white background for chromosome delineation}

\item{...}{Parameters passed onto the actual plot command}
}
\description{
Displays the read-depth array in a linear fashion. Can report CNV as all gray or use color calls.
If sampleId is provided, title annotations will be added.
}
\examples{
sampleId='TCGA-14-1402-02A_ds'
inputDir <- system.file('extdata', package = "BACDAC")
outputDir=tempdir()
segmentationFile = file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
segmentation= loadSegmentationFile(segmentationFile) # chr, start, end, rd per segment
thirtyKbFile=file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds'))
readDepthBinnedData = readRDS(file=thirtyKbFile )
  yAxisLimits = starCloudResult$plotAxisLimits$nrdAxisLims
op <- par(mfrow=c(3,1),mai=c(.25,0.5, 0.3,0.25), mgp=c(2, .5, 0))
# default cnv color coding and annotations
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=sampleId,segmentation=segmentation)
# example with no cnv color coding
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=sampleId,noDelAmpDetection=TRUE,segmentation=segmentation)
# reviewer preferred cnv color coding
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=NULL,segmentation=segmentation, gainColor = 'red', lossColor= 'blue' )

}
