% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGenomePlot.R
\name{linearGenomePlot}
\alias{linearGenomePlot}
\title{Draw a linear genome plot}
\usage{
linearGenomePlot(
  readDepthBinnedData,
  readDepthBinSize = 30000,
  yLimQuantile = 0.99,
  sampleId = NULL,
  alternateId = NULL,
  segmentation = NULL,
  allelicSegments = NULL,
  noDelAmpDetection = FALSE,
  gainColor = "blue",
  lossColor = "red",
  zebraStrips = FALSE,
  ...
)
}
\arguments{
\item{readDepthBinnedData}{list of two arrays: one with (normalized) read count per each window, second with linear genome position of each window (masked windows have been removed)}

\item{readDepthBinSize}{size of each window}

\item{yLimQuantile}{To prevent an outlier from dominating the plot, we set the max y axis to this quantile (number 0-1, 1=100\%, 0.5=median)}

\item{sampleId}{plot y axis with normalize read depth to 2 for the median frequency}

\item{segmentation}{read depth data.frame with required columns: chr, start, end, rd; optional: cnvState for color coded linear genome plot}

\item{allelicSegments}{segments used in \code{calculatePloidy} then augmented with major and minor allele specific copy number}

\item{noDelAmpDetection}{do not color code deletions and gains in genome plot}

\item{gainColor}{color to use for gains in linear genome plot, default is blue}

\item{lossColor}{color to use for losses in linear genome plot, default is red}

\item{zebraStrips}{option to have alternating gray/white background for chromosome delineation}

\item{...}{Parameters passed onto the actual plot command}
}
\description{
Displays the read-depth array in a linear fashion. Can report CNV as all gray or use color calls.
If sampleId is provided, title annotations will be added.
}
\examples{
sampleId='TCGA-14-1402-02A_ds'
inputDir='/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Rprojects/BACDAC/inst/extdata'
outputDir='/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Routput/BACDAC'
segmentationFile = file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
segmentation <- read.csv(segmentationFile,comment.char = '#', header = TRUE); dim(segmentation)
thirtyKbFile=file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds'))
readDepthBinnedData = readRDS(file=thirtyKbFile )

op <- par(mfrow=c(3,1),mai=c(.25,0.5, 0.3,0.25), mgp=c(2, .5, 0))
# default cnv color coding and annotations
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=sampleId,segmentation=segmentation)
# example with no cnv color coding
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=sampleId,noDelAmpDetection=TRUE,segmentation=segmentation)
# reviewer preferred cnv color coding
linearGenomePlot(readDepthBinnedData=readDepthBinnedData,sampleId=NULL,segmentation=segmentation, gainColor = 'red', lossColor= 'blue' )

}
