% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationPlot.R
\name{loadStarsInTheClouds}
\alias{loadStarsInTheClouds}
\title{load all the data necessary to make the constellation plot}
\usage{
loadStarsInTheClouds(
  sampleId,
  inputDir,
  readDepthPer30kbBin,
  hetScorePerBinFile,
  hsNormMat,
  testVals,
  wsz = 30000,
  mainPeakNRD,
  expReadsIn2NPeak_1bp
)
}
\arguments{
\item{sampleId}{sample Identifier}

\item{readDepthPer30kbBin}{read depth for 30 kb bins, with bin index in linear coordinates}

\item{hsNormMat}{heterozygosity score mask, used to look for places in the 23 TCGA normals where more than half dropped below the a (i.e. 0.975) cutoff.}

\item{testVals}{used to find each possible heterozygosity value for each copy number level (find the right spots for the stars)}

\item{mainPeakNRD}{the NRD of the main Peak in the cnvBinned data.  Will equal 2 if main peak is the normal peak}

\item{expReadsIn2NPeak_1bp}{expected number of reads in a 1 bp bin for the diploid peak}
}
\description{
load all the data necessary to make the constellation plot
}
\examples{
\dontrun{

  library(BACDAC)
  library(logging)

  # intialize
  starCloudResult=NULL;
  starCloudPlotInputs=NULL
  saveHetScoreRdata=FALSE
  noPdf=TRUE

  ## load two reference files  ---------------
  # hsNormMat/lohMat: LOH analysis mask, used to look for places in 23 TCGA normals where more than half dropped below the a (i.e. 0.975) cutoff.
  # testVals: used to find each possible heterozygosity value for each copy number level (find the right spots for the stars)
  hsNormMat <- bmdTools::loadRdata('/research/labs/experpath/vasm/shared/NextGen/Misc/pipelineInputs/hetScoreAnalysis/lohMat.Rdata') # aka lohMat
  testVals <-  bmdTools::loadRdata(file.path('/research/labs/experpath/vasm/shared/NextGen/Misc/pipelineInputs/hetScoreAnalysis/testVals.Rdata'))

  sampleId='TCGA-14-1402-02A_ds'; alternateId=66301

  # directory with input files:
  inputDir <- system.file('extdata', package = "BACDAC") # or '/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Rprojects/BACDAC/inst/extdata'
  readDepthPer30kbBin=  readRDS(file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds')))

  # directory for output files:
  outputDir='/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Routput/BACDAC'

  # result from heterozgygosityScore
  hetScorePerBinWigFile <- file.path(outputDir, 'reports', paste0(sampleId, '_hetScorePerBin.wig.gz'))

  # result from calculatePloidy
  calcPloidyResultOutputFile=file.path(outputDir, paste0(sampleId, '_calculatePloidyResult.Rds'))
  calcPloidyResult = readRDS(calcPloidyResultOutputFile)
  mainPeakNRD=getMainPeakNRD(calcPloidyResult)
  expReadsIn2NPeak_1bp=calcPloidyResult$expReadsIn2NPeak_1bp

  ### get the needed input values for the plot
  if(is.null(starCloudPlotInputs)){     # run once only... it takes about 3-5 minutes
    starCloudPlotInputs=loadStarsInTheClouds(sampleId, inputDir, readDepthPer30kbBin,hetScorePerBinWigFile, hsNormMat, testVals, mainPeakNRD=mainPeakNRD, expReadsIn2NPeak_1bp=expReadsIn2NPeak_1bp)
  }

  ### make the plot
  if (!noPdf) {
    constellationPdfFile <- getTypedFile('constellationPlot',dir = outputDir, list(sampleId = sampleId))
    pdf(file = constellationPdfFile@path, paper="a4r", width=8, height=10, title=paste0('hetScVsCN_',sampleId))
    loginfo('writing pdf: \%s', constellationPdfFile@path)
    on.exit(dev.off(), add=TRUE)
  }
  op <- par(mfrow=c(1,1),mar=c(5,4,3.5,3.5),mgp=c(1.5, 0.5,0))
  starCloudResult=plotStarsInTheClouds(sampleId, alternateId,starCloudPlotInputs, diploidPeakNRD=NULL, tau=min(1,calcPloidyResult$percentTumor/100),
                                       plotEachChrom=FALSE, mainPeakNRD=mainPeakNRD,
                                       segmentData=calcPloidyResult$segmentData, peakInfo=calcPloidyResult$peakInfo,
                                       digitalPeakZone =calcPloidyResult[['iterationStatsAll']][['digitalPeakZone']],
                                       paperMode=TRUE)
  par(op)



  ##### TWO PLOT OPTION ######
  ### plot constellation plot with linear plot ----
  # load read depth data
  thirtyKbFile=file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds'))
  readDepthPer30kbBin = readRDS(file=thirtyKbFile )
  # load segmentation data
  segmentationFile <- file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
  segmentation= loadSegmentationFile(segmentationFile)
  hetScorePerBinWigFile <- file.path(outputDir, 'reports', paste0(sampleId, '_hetScorePerBin.wig.gz'))
  hetScore <- loadHetScoreFromWig(hetScorePerBinWigFile)



  layout( matrix(c(1,2,2),
                 nrow=1),
          heights= c(1),
          widths = c(1.5,2))   # Widths of the two columns
  layout.show(2)

  # left figure
  op <- par(mar=c(5,3, 2,2.5),mgp=c(1.5, 0.5,0))
  starCloudResult=plotStarsInTheClouds(sampleId=NULL, alternateId=NULL,starCloudPlotInputs, diploidPeakNRD=NULL, tau=min(1,calcPloidyResult$percentTumor/100),
                                       plotEachChrom=FALSE, mainPeakNRD=mainPeakNRD,
                                       segmentData=calcPloidyResult$segmentData, peakInfo=calcPloidyResult$peakInfo,
                                       digitalPeakZone =calcPloidyResult[['iterationStatsAll']][['digitalPeakZone']],
                                       paperMode=FALSE,plotCex=1.3)
  par(op)

  # right figure
  op <- par(mar=c(5, 3, 2,1),mgp=c(1.5, 0.5,0))
  linearGenomePlot( readDepthBinnedData=readDepthPer30kbBin, wsz=readDepthPer30kbBin$windowSize, segmentation=segmentation,
                    allelicSegments=starCloudResult$allelicSegments,
                    gainColor = 'blue', lossColor= 'red', yAxisLimits = yAxisLimits)
  par(op)


  ### save the results for future inspection ----
  if(saveHetScoreRdata){
    starLookUp=makeStarLookUpTable(starCloudResult,percentTumor=calcPloidyResult$percentTumor)

    # TODO: maybe this should be moved inside plotStarsInTheClouds?
    starCloudData <- list(
      #outputs:
      starLookUp=starLookUp,
      expReadsIn2NPeak_1bp=expReadsIn2NPeak_1bp,
      percentTumor=calcPloidyResult$percentTumor,
      ploidyCN=starCloudResult$ploidyCN,
      diploidPeakNRD=starCloudResult$diploidPeakNRD,
      mainPeakNRD=mainPeakNRD,
      ploidySegments=calcPloidyResult$segmentData,
      lohContent=starCloudResult$lohContent,
      plotAxisLimits=starCloudResult$plotAxisLimits,
      # inputs:
      forceFirstDigPeakCopyNum=forceFirstDigPeakCopyNum,
      grabDataPercentManual=grabDataPercentManual,
      minPeriodManual=minPeriodManual,
      origMaxPercentCutoffManual=origMaxPercentCutoffManual,
      allowedTumorPercent=allowedTumorPercent
    )

    starCloudResultFile=file.path(outputDir, paste0(sampleId, '_StarCloudResult.Rds'))
    loginfo('saving result to: \%s',starCloudResultFile)
    saveRDS(starCloudData, file=starCloudResultFile)
  }


}
}
