% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{makeHetScoreReportPdf}
\alias{makeHetScoreReportPdf}
\title{Make 3-subplot hetScore analysis PDF}
\usage{
makeHetScoreReportPdf(
  hetScorePerBinFile,
  hetScorePerArmFile,
  segmentation = NULL,
  readDepthPer30kbBin = NULL,
  readDepthBinSize = 30000,
  sampleId,
  alternateId = NULL,
  outputDir,
  gainColor = "blue",
  lossColor = "red",
  noDelGainCalls = FALSE,
  noPdf = FALSE
)
}
\arguments{
\item{hetScorePerBinFile}{full path to Heterozygosity Score per bin wig file
as created by \code{calculateHetScore}}

\item{hetScorePerArmFile}{full path to Heterozygosity Score per arm csv file
as created by \code{calculateHetScore}}

\item{segmentation}{identified regions of the genome with constant read depth. Data.frame with required columns:
chr, start, end, rd; optional: cnvState (1=loss, 2=normal, 3=gain) for color coded linear linear genome plot}

\item{readDepthPer30kbBin}{list with two equal length arrays:
\itemize{
\item \code{readDepthArray} read depth for 30 kb bins, normalized for GC content and other artifacts
\item \code{goodWindowArray} linear genome position of each window (bin), masked windows have been removed
}}

\item{readDepthBinSize}{bp size of the bins in the read depth array, expecting 30000 or 100000}

\item{sampleId}{sample Identifier}

\item{alternateId}{optional secondary sample identifier}

\item{outputDir}{full path to directory to output all files and intermediate objects created}

\item{gainColor}{color to use for gains in linear genome plot, default is blue}

\item{lossColor}{color to use for losses in linear geneome plot, default is red}

\item{noDelGainCalls}{do not color code deletions and gains in genome plot}

\item{noPdf}{if present (TRUE), pdf files will not be generated, instead plots are drawn on default device}
}
\description{
Will show 3 separate plots aligning segmentation, heterozygosity score per 30K and
heterozygosity score per arm.
}
\details{
Will print to plot window if \code{noPdf=TRUE}. Use \code{noPdf=FALSE} and provide an \code{outputDir} to save
to file.
}
\examples{
library(BACDAC)
sampleId='TCGA-14-1402-02A_ds'

# inputDir is the path to input data, here we use the example data provided in the package.
inputDir <- system.file('extdata', package = "BACDAC")
hetScorePerArmFile <- file.path(inputDir, paste0(sampleId, '_hetScorePerArm.csv'))
hetScorePerBinFile <- file.path(inputDir, paste0(sampleId, '_hetScorePerBin.wig.gz'))

segmentationFile <- file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
segmentation <- read.csv(segmentationFile, comment.char = '#') # chr, start, end, rd per
segmentation <- checkSegmentation(segmentation)
thirtyKbFile <- file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds'))
readDepthPer30kbBin <- readRDS(file=thirtyKbFile )


# plot to window using default segmentation coloring
makeHetScoreReportPdf(
   hetScorePerBinFile=hetScorePerBinFile,hetScorePerArmFile=hetScorePerArmFile,
   segmentation=segmentation,readDepthPer30kbBin=readDepthPer30kbBin,
   readDepthBinSize=readDepthPer30kbBin$windowSize,sampleId=sampleId,
   noPdf=TRUE, outputDir=NULL)

}
