% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{makeHetScoreReportPdf}
\alias{makeHetScoreReportPdf}
\title{Make 3-subplot hetScore analysis PDF}
\usage{
makeHetScoreReportPdf(
  hetScorePerBinWigFile,
  hetScorePerArmFile,
  segmentation = NULL,
  readDepthBinnedData = NULL,
  readDepthBinSize = 30000,
  sampleId,
  alternateId = NULL,
  outputDir,
  gainColor = "blue",
  lossColor = "red",
  noDelAmpDetection = FALSE,
  noPdf = FALSE
)
}
\arguments{
\item{hetScorePerBinWigFile}{full path to Heterozygosity Score per bin wig file as created by \code{calculateHetScore}}

\item{hetScorePerArmFile}{full path to Heterozygosity Score per arm csv file as created by \code{calculateHetScore}}

\item{segmentation}{read depth data.frame with required columns: chr, start, end, rd; optional: cnvState for color coded linear genome plot}

\item{readDepthBinnedData}{list of two arrays: one with (normalized) read count per each window, second with linear genome position of each window (masked windows have been removed)}

\item{readDepthBinSize}{size of each window}

\item{sampleId}{sample Identifier}

\item{alternateId}{optional secondary sample identifier}

\item{outputDir}{full path for output files}

\item{gainColor}{color to use for gains in linear genome plot, default is blue}

\item{lossColor}{color to use for losses in linear geneome plot, default is red}

\item{noDelAmpDetection}{do not color code deletions and gains in genome plot}

\item{noPdf}{if present, do not create pdfs}
}
\description{
Will show 3 separate plots contrasting CNV, heterozygosity score per 30K and heterozygosity score per arm.
}
\examples{
sampleId='TCGA-14-1402-02A_ds'
alternateId=NULL
outputDir <- tempdir()

# inputDir is the path to input data, either the users data, or in this example, the package data
inputDir <- system.file('extdata', package = "BACDAC")
outputDir = '/research/labs/experpath/vasm/shared/NextGen/johnsonsh/Routput/BACDAC'
hetScoreDir=file.path(outputDir, 'reports') # outputDir here is the outputDir for calculateHetScore.R
hetScorePerArmFile <- file.path(hetScoreDir, paste0(sampleId, '_hetScorePerArm.csv'))
hetScorePerBinWigFile <- file.path(hetScoreDir, paste0(sampleId, '_hetScorePerBin.wig.gz'))

segmentationFile = file.path(inputDir, paste0(sampleId, '_segmentation.csv'))
segmentation <- read.csv(segmentationFile,comment.char = '#', header = TRUE); dim(segmentation)
thirtyKbFile=file.path(inputDir, paste0(sampleId,'_','readDepthPer30kbBin.Rds'))
readDepthBinnedData = readRDS(file=thirtyKbFile )

gainColor = 'blue'; lossColor= 'red'; noDelAmpDetection=FALSE; noPdf=TRUE

 op <- par(mfrow=c(3,1),mai=c(.25,0.5, 0.3,0.25), mgp=c(2, .5, 0))
# default cnv coloring (by default) and annotations
makeHetScoreReportPdf(
   hetScorePerBinWigFile=hetScorePerBinWigFile,hetScorePerArmFile=hetScorePerArmFile,
   segmentation=segmentation,readDepthBinnedData=readDepthBinnedData,
   readDepthBinSize=readDepthBinnedData$windowSize,sampleId=sampleId,noPdf=TRUE)
}
