% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{peaksByDensity}
\alias{peaksByDensity}
\title{Use density on 100 kb binned read depth to determine read depth distribution (peaks)}
\usage{
peaksByDensity(
  sampleId,
  readDepthPer100kbBin,
  segmentation,
  segmentationBinSize = 30000,
  wszPeaks = 1e+05,
  grabDataPercentManual = -1,
  origMaxPercentCutoffManual = -1,
  addAreaLinesToPlot = FALSE,
  qualityPostNorm = NULL,
  omitAnnotations = FALSE,
  alternateId = NULL
)
}
\arguments{
\item{sampleId}{sample Identifier}

\item{readDepthPer100kbBin}{read depth for 100kb bins, with bin index in linear coordinates}

\item{segmentation}{identified regions of the genome with constant read depth}

\item{segmentationBinSize}{bin size used for the read depth in the segmentation data}

\item{wszPeaks}{window size for peaks by density}

\item{grabDataPercentManual}{portion of main peak data to grab, other peaks will be scaled based on read depth (x location) set to -1 to use main Peak width}

\item{origMaxPercentCutoffManual}{peaks smaller than this portion of the max peak are not considered; set to -1 to use default value}

\item{qualityPostNorm}{used to set max allowed value for grabDataPercent}

\item{alternateId}{secondary sample identifier}
}
\description{
peaks are returned in rank order of read depth with highest frequency read depth peak listed first
}
\keyword{internal}
