% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyUtils.R
\name{pkmodToNRD}
\alias{pkmodToNRD}
\title{as calculated in constellation plot
converts pkmod rather than rd to nrd, normalized read depth}
\usage{
pkmodToNRD(segmentData, peakInfo, rdNormX_2Npeak)
}
\description{
as calculated in constellation plot
converts pkmod rather than rd to nrd, normalized read depth
}
