% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{plotEmptyLinearGenomePlot}
\alias{plotEmptyLinearGenomePlot}
\title{create a plot to fill the empty space for now}
\usage{
plotEmptyLinearGenomePlot(chromsToPlot)
}
\arguments{
\item{chromsToPlot}{Vector of chromosome numbers to plot}
}
\description{
create a plot to fill the empty space for now
}
