% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{plotEmptyLinearGenomePlot}
\alias{plotEmptyLinearGenomePlot}
\title{create a plot to fill the empty space for now}
\usage{
plotEmptyLinearGenomePlot(chromsToPlot, coords)
}
\arguments{
\item{chromsToPlot}{vector of chromosome (as integers) to plot}

\item{coords}{object describing linear coordinate space for the chromosomes}
}
\description{
create a plot to fill the empty space for now
}
