% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosityScore.R
\name{plotHetScorePerBin}
\alias{plotHetScorePerBin}
\title{Plot the binned heterozygosity score}
\usage{
plotHetScorePerBin(
  hetScore,
  sampleId = NULL,
  yMap = function(y) {     y },
  ylab = "Heterozygosity Score",
  allelicSegments = NULL,
  zebraStrips = FALSE
)
}
\arguments{
\item{hetScore}{binned heterozygosity scores as loaded from the hetScore wig file}

\item{sampleId}{sample id, will be used as the prefix for all input and output file names}

\item{yMap}{A function that turns the actual y value into a position on screen, transform y coordinates for drawing purposes}

\item{ylab}{label for y axis default "Heterozygosity Score"}

\item{allelicSegments}{segmentData from \code{calculatePloidy} then augmented in \code{plotStarsInTheClouds} with major_copy_number and minor_copy_number}

\item{zebraStrips}{option to have alternating gray/white background for chromosome delineation}
}
\description{
These plots show the heterozygosity scores per bin.
For per arm values, see \link{plotHetScorePerArm}. \code{sampleId} is optional,
for plot annotation only.
}
\seealso{
Other hetScore: 
\code{\link{calculateHetScoreTestStatisticPerCoverage}()},
\code{\link{makeAndSaveHetScorePerArm}()},
\code{\link{plotHetScorePerArm}()},
\code{\link{saveHetScoreToWig}()}
}
\concept{hetScore}
