% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationPlot.R
\name{twoPanelReport}
\alias{twoPanelReport}
\title{draw the constellation plot and the linear genome plot side by side}
\usage{
twoPanelReport(
  starCloudPlotInputs,
  calcPloidyResult,
  readDepthPer30kbBin,
  segmentation,
  sampleId = NULL,
  alternateId = NULL,
  diploidPeakNRD = NULL,
  gainColor = "blue",
  lossColor = "red",
  leftFigLabel = NULL,
  rightFigLabel = NULL
)
}
\description{
draw the constellation plot and the linear genome plot side by side
}
